requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
streamlit
openai
python-dotenv
requests
streamlit-aggrid
